// src/swagger/share/models/share-dto.ts
/* tslint:disable */
/* eslint-disable */
/**
 * REST API SPEC FOR MOVESONG FRONTEND
 * REST API SPEC FOR MOVESONG FRONTEND
 *
 * OpenAPI spec version: v1
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
/**
 * 
 * @export
 * @interface ShareDto
 */
export interface ShareDto {
    /**
     * 
     * @type {number}
     * @memberof ShareDto
     */
    id?: number;
    /**
     * 
     * @type {string}
     * @memberof ShareDto
     */
    playlistId?: string;
    /**
     * 
     * @type {string}
     * @memberof ShareDto
     */
    sharedPlaylistName?: string;
    /**
     * 
     * @type {string}
     * @memberof ShareDto
     */
    movesongEmail?: string;
    /**
     * 
     * @type {boolean}
     * @memberof ShareDto
     */
    visible?: boolean;
    /**
     * 
     * @type {number}
     * @memberof ShareDto
     */
    views?: number;
    /**
     * 
     * @type {string}
     * @memberof ShareDto
     */
    sharePlatformType?: string;
    /**
     * 
     * @type {number}
     * @memberof ShareDto
     */
    selectedBackgroundId?: number;
    /**
     * 
     * @type {string}
     * @memberof ShareDto
     */
    thumbnailUrl?: string;
}